Extract Arcs proto decoding from ir::TagClaim.
Part of the ongoing effort to detangle Arcs protos from the Raksha IR
(issue #118).
