+++ src/ir/proto/BUILD
@@ -67,6 +67,17 @@ cc_library(
     ],
 )
 
+cc_library(
+    name = "tag_claim",
+    srcs = ["tag_claim.cc"],
+    hdrs = ["tag_claim.h"],
+    deps = [
+        "//src/ir",
+        "//src/ir:access_path",
+        "//third_party/arcs/proto:manifest_cc_proto",
+    ],
+)
+
 cc_test(
     name = "access_path_test",
     srcs = ["access_path_test.cc"],
